--- src/main/java/no/digdir/oidc/testclient/service/OIDCIntegrationException.java
@@ -1,9 +1,0 @@
-package no.digdir.oidc.testclient.service;
-
-public class OIDCIntegrationException extends RuntimeException {
-
-    public OIDCIntegrationException(String message) {
-        super(message);
-    }
-
-}
