--- src/main/java/no/digdir/oidc/testclient/service/ProtocolInteraction.java
@@ -1,14 +1,0 @@
-package no.digdir.oidc.testclient.service;
-
-import lombok.Builder;
-import lombok.Data;
-
-@Builder
-@Data
-public class ProtocolInteraction {
-
-    private String id;
-    private String text;
-    private String interaction;
-
-}
