+++ src/main/java/no/idporten/tools/oidc/democlient/config/FeatureSwichProperties.java
@@ -1,0 +1,16 @@
+package no.idporten.tools.oidc.democlient.config;
+
+import lombok.Data;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.validation.annotation.Validated;
+
+@Configuration
+@Data
+@Validated
+@ConfigurationProperties(prefix = "oidc-demo-client.features")
+public class FeatureSwichProperties {
+
+    private boolean authorizationDetailsEnabled = false;
+
+}
