+++ src/main/java/no/idporten/tools/oidc/democlient/service/LoggingResourceRetriever.java
@@ -1,0 +1,31 @@
+package no.idporten.tools.oidc.democlient.service;
+
+import com.nimbusds.jose.util.DefaultResourceRetriever;
+import com.nimbusds.jose.util.Resource;
+import lombok.extern.slf4j.Slf4j;
+
+import java.io.IOException;
+import java.net.URL;
+
+@Slf4j
+public class LoggingResourceRetriever extends DefaultResourceRetriever {
+
+    public LoggingResourceRetriever(int connectTimeout, int readTimeout) {
+        super(connectTimeout, readTimeout);
+    }
+
+    @Override
+    public Resource retrieveResource(URL url) throws IOException {
+        try {
+            Resource resource = super.retrieveResource(url);
+            if (log.isInfoEnabled()) {
+                log.info("Loaded resource from {}", url);
+            }
+            return resource;
+        } catch (IOException e) {
+            log.info("Failed to load resource from {}", url, e);
+            throw e;
+        }
+    }
+
+}
