+++ src/main/java/no/idporten/tools/oidc/democlient/service/OIDCIntegrationException.java
@@ -1,0 +1,9 @@
+package no.idporten.tools.oidc.democlient.service;
+
+public class OIDCIntegrationException extends RuntimeException {
+
+    public OIDCIntegrationException(String message) {
+        super(message);
+    }
+
+}
