+++ src/main/java/no/idporten/tools/oidc/democlient/service/ProtocolInteraction.java
@@ -1,0 +1,14 @@
+package no.idporten.tools.oidc.democlient.service;
+
+import lombok.Builder;
+import lombok.Data;
+
+@Builder
+@Data
+public class ProtocolInteraction {
+
+    private String id;
+    private String text;
+    private String interaction;
+
+}
